bug 1302459: Remove translate_locales_name
The current code marks locale names as translated strings through an
unused template file that is generated by running:

  ./manage.py translate_locales_names

Because these are now in the generated settings file languages.py, we
can eliminate some duplication in generated files.

When locales are updated, existing translations will still be in
django.po, but associated with a different source file.